refactor to pages directory
