+++ src/pages/Launches.js
@@ -1,6 +1,6 @@
 import React from 'react';
 import { withRouter } from 'react-router-dom';
-import CardList from './CardList';
+import CardList from '../containers/CardList';
 
 const Launches = () => (
   <div>
